+++ src/KubeLibrary/KubeLibrary.py
@@ -324,6 +324,23 @@ class KubeLibrary(object):
         deployments = [item for item in ret.items if r.match(item.metadata.name)]
         return deployments
 
+    def get_replicasets_in_namespace(self, name_pattern, namespace, label_selector=""):
+        """Gets replicasets matching pattern in given namespace.
+
+        Can be optionally filtered by label. e.g. label_selector=label_key=label_value
+
+        Returns list of  replicasets.
+
+        - ``name_pattern``:
+          replicaset name pattern to check
+        - ``namespace``:
+          Namespace to check
+        """
+        ret = self.appsv1.list_namespaced_replica_set(namespace, watch=False, label_selector=label_selector)
+        r = re.compile(name_pattern)
+        replicasets = [item for item in ret.items if r.match(item.metadata.name)]
+        return replicasets
+
     def get_jobs_in_namespace(self, name_pattern, namespace, label_selector=""):
         """Gets jobs matching pattern in given namespace.
 
@@ -365,6 +382,14 @@ class KubeLibrary(object):
           List of deployments objects
         """
         return [d.metadata.name for d in deployments]
+
+    def filter_replicasets_names(self, replicasets):
+        """Filter replicaset  names for list of replicasets .
+        Returns list of strings.
+        - ``replicasets``:
+          List of replicasets objects
+        """
+        return [d.metadata.name for d in replicasets]
 
     def filter_pods_names(self, pods):
         """Filter pod names for list of pods.